requirements.txt
PySide6>=6.8.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-jose==3.3.0
python-dotenv==1.0.0
loguru==0.7.2
pandas==2.1.1
reportlab==4.0.4
pillow==10.0.0
cryptography==41.0.4
